+++ CHANGELOG.md
@@ -1,60 +1,152 @@
-## [1.3.1](https://github.com/devperimental/component-api-extensions/compare/v1.3.0...v1.3.1) (2023-03-20)
+# [1.4.0-alpha-5.14](https://github.com/devperimental/component-api-extensions/compare/v1.4.0-alpha-5.13...v1.4.0-alpha-5.14) (2024-07-14)
 
+
+### Features
+
+* update serilog setup ([b1e4335](https://github.com/devperimental/component-api-extensions/commit/b1e4335e4a1f2e3eb180d6d196fc810c393caadb))
+
+# [1.4.0-alpha-5.13](https://github.com/devperimental/component-api-extensions/compare/v1.4.0-alpha-5.12...v1.4.0-alpha-5.13) (2024-03-12)
+
+
+### Features
+
+* add header info ([5b4bd8a](https://github.com/devperimental/component-api-extensions/commit/5b4bd8a4bad5ad41d8d1a8e5b8535b03d1864a01))
+
+# [1.4.0-alpha-5.12](https://github.com/devperimental/component-api-extensions/compare/v1.4.0-alpha-5.11...v1.4.0-alpha-5.12) (2024-03-12)
+
+
+### Features
+
+* add sec ([88b4962](https://github.com/devperimental/component-api-extensions/commit/88b4962cc736088f306e21bb2835551dadd0a373))
+
+# [1.4.0-alpha-5.11](https://github.com/devperimental/component-api-extensions/compare/v1.4.0-alpha-5.10...v1.4.0-alpha-5.11) (2024-03-12)
+
+
+### Features
+
+* downgrade to 6.4 for .net 6 ([a701cb1](https://github.com/devperimental/component-api-extensions/commit/a701cb16976912618687acba82f85cd25e4f4965))
+
+# [1.4.0-alpha-5.10](https://github.com/devperimental/component-api-extensions/compare/v1.4.0-alpha-5.9...v1.4.0-alpha-5.10) (2023-11-21)
+
+
+### Features
+
+* update ASPNETCORE_ENVIRONMENT ([0d324cc](https://github.com/devperimental/component-api-extensions/commit/0d324ccf1af7899606c9e7159bc1816904b1762e))
+
+# [1.4.0-alpha-5.9](https://github.com/devperimental/component-api-extensions/compare/v1.4.0-alpha-5.8...v1.4.0-alpha-5.9) (2023-11-18)
+
+
+### Features
+
+* package updates ([c12b8f7](https://github.com/devperimental/component-api-extensions/commit/c12b8f7b98077aa783a4ee65ef2c77e54a21ccfd))
+
+# [1.4.0-alpha-5.8](https://github.com/devperimental/component-api-extensions/compare/v1.4.0-alpha-5.7...v1.4.0-alpha-5.8) (2023-11-18)
+
+
+### Features
+
+* downgrade to .net60 ([58bb26b](https://github.com/devperimental/component-api-extensions/commit/58bb26b643ba33681f9c0a4c6304764be67bef4c))
+
+# [1.4.0-alpha-5.7](https://github.com/devperimental/component-api-extensions/compare/v1.4.0-alpha-5.6...v1.4.0-alpha-5.7) (2023-09-15)
+
+
+### Features
+
+* Tweak logging setup ([cd37612](https://github.com/devperimental/component-api-extensions/commit/cd376127a178115145cc86fefe75f804f82f2638))
+
+# [1.4.0-alpha-5.6](https://github.com/devperimental/component-api-extensions/compare/v1.4.0-alpha-5.5...v1.4.0-alpha-5.6) (2023-09-15)
+
+
+### Features
+
+* add local cloudwatch logging ([bca13ce](https://github.com/devperimental/component-api-extensions/commit/bca13ceabee071f318664b59854c3fbd00cff8d1))
+
+# [1.4.0-alpha-5.5](https://github.com/devperimental/component-api-extensions/compare/v1.4.0-alpha-5.4...v1.4.0-alpha-5.5) (2023-09-13)
+
+
+### Features
+
+* updates ([9cb3c2e](https://github.com/devperimental/component-api-extensions/commit/9cb3c2e91d4abbedb866897b1860d5f30ee8257b))
+
+# [1.4.0-alpha-5.4](https://github.com/devperimental/component-api-extensions/compare/v1.4.0-alpha-5.3...v1.4.0-alpha-5.4) (2023-06-04)
+
+
+### Features
+
+* refresh to v7 ([c3c96c5](https://github.com/devperimental/component-api-extensions/commit/c3c96c564e4bbad8600fdeae1c56ef68a80c6896))
+
+# [1.4.0-alpha-5.3](https://github.com/devperimental/component-api-extensions/compare/v1.4.0-alpha-5.2...v1.4.0-alpha-5.3) (2023-06-01)
+
+
+### Features
+
+* Tweak ([3cbbfce](https://github.com/devperimental/component-api-extensions/commit/3cbbfcef3d6adcdd8bc9ece223a39d7e32a984ca))
+
+# [1.4.0-alpha-5.2](https://github.com/devperimental/component-api-extensions/compare/v1.4.0-alpha-5.1...v1.4.0-alpha-5.2) (2023-06-01)
+
+
+### Features
+
+* fix namespace ([4aa4110](https://github.com/devperimental/component-api-extensions/commit/4aa4110f091a822ffa292f244666475828b2863a))
 
 ### Bug Fixes
 
-* refresh versions ([8d876da](https://github.com/devperimental/component-api-extensions/commit/8d876daa0efeda46ca0e87bf4c1c19bb0b917a27))
+- refresh versions ([8d876da](https://github.com/devperimental/component-api-extensions/commit/8d876daa0efeda46ca0e87bf4c1c19bb0b917a27))
+
+# [1.4.0-alpha-5.1](https://github.com/devperimental/component-api-extensions/compare/v1.3.0...v1.4.0-alpha-5.1) (2023-04-05)
+
+### Bug Fixes
+
+- refresh versions ([8d876da](https://github.com/devperimental/component-api-extensions/commit/8d876daa0efeda46ca0e87bf4c1c19bb0b917a27))
+
+## [1.3.1](https://github.com/devperimental/component-api-extensions/compare/v1.3.0...v1.3.1) (2023-03-20)
+
+### Bug Fixes
+
+- refresh versions ([8d876da](https://github.com/devperimental/component-api-extensions/commit/8d876daa0efeda46ca0e87bf4c1c19bb0b917a27))
 
 ## [1.3.1-alpha-5.1](https://github.com/devperimental/component-api-extensions/compare/v1.3.0...v1.3.1-alpha-5.1) (2023-03-20)
 
+### Features
 
-### Bug Fixes
-
-* refresh versions ([8d876da](https://github.com/devperimental/component-api-extensions/commit/8d876daa0efeda46ca0e87bf4c1c19bb0b917a27))
+- Add storage DI ([be1d936](https://github.com/devperimental/component-api-extensions/commit/be1d9367caa6c433be3adcc118bd1a4e899576ed))
 
 # [1.3.0](https://github.com/devperimental/component-api-extensions/compare/v1.2.0...v1.3.0) (2023-03-19)
 
-
 ### Features
 
-* updates to libs ([51ab3ff](https://github.com/devperimental/component-api-extensions/commit/51ab3ffc2de8a611019e46abb89bd2860208f7ea))
+- updates to libs ([51ab3ff](https://github.com/devperimental/component-api-extensions/commit/51ab3ffc2de8a611019e46abb89bd2860208f7ea))
 
 # [1.3.0-alpha-4.1](https://github.com/devperimental/component-api-extensions/compare/v1.2.0...v1.3.0-alpha-4.1) (2023-03-19)
 
-
 ### Features
 
-* updates to libs ([51ab3ff](https://github.com/devperimental/component-api-extensions/commit/51ab3ffc2de8a611019e46abb89bd2860208f7ea))
+- updates to libs ([51ab3ff](https://github.com/devperimental/component-api-extensions/commit/51ab3ffc2de8a611019e46abb89bd2860208f7ea))
 
 # [1.2.0](https://github.com/devperimental/component-api-extensions/compare/v1.1.0...v1.2.0) (2023-03-17)
 
-
 ### Bug Fixes
 
-* merge conflict ([e577049](https://github.com/devperimental/component-api-extensions/commit/e57704967fc48d6d93cbd8e11290ca951d47e100))
-
+- merge conflict ([e577049](https://github.com/devperimental/component-api-extensions/commit/e57704967fc48d6d93cbd8e11290ca951d47e100))
 
 ### Features
 
-* update to latest version ([62c217f](https://github.com/devperimental/component-api-extensions/commit/62c217fbe357536f23e4e7dd7e38d9c22506d601))
+- update to latest version ([62c217f](https://github.com/devperimental/component-api-extensions/commit/62c217fbe357536f23e4e7dd7e38d9c22506d601))
 
 # [1.1.0](https://github.com/devperimental/component-api-extensions/compare/v1.0.0...v1.1.0) (2023-03-16)
 
-
 ### Features
 
-* add secrets support ([628b2e6](https://github.com/devperimental/component-api-extensions/commit/628b2e63d58b139606ae440e8232b4f91f2304bb))
+- add secrets support ([628b2e6](https://github.com/devperimental/component-api-extensions/commit/628b2e63d58b139606ae440e8232b4f91f2304bb))
 
 # [1.1.0-alpha-2.1](https://github.com/devperimental/component-api-extensions/compare/v1.0.0...v1.1.0-alpha-2.1) (2023-03-16)
 
-
 ### Features
 
-* add secrets support ([628b2e6](https://github.com/devperimental/component-api-extensions/commit/628b2e63d58b139606ae440e8232b4f91f2304bb))
+- add secrets support ([628b2e6](https://github.com/devperimental/component-api-extensions/commit/628b2e63d58b139606ae440e8232b4f91f2304bb))
 
 # 1.0.0 (2023-03-04)
 
-
 ### Features
 
-* initial commit ([36fe85b](https://github.com/devperimental/component-api-extensions/commit/36fe85bd4b0a6cc17fb49e7906a0cf25be7deb1c))
+- initial commit ([36fe85b](https://github.com/devperimental/component-api-extensions/commit/36fe85bd4b0a6cc17fb49e7906a0cf25be7deb1c))